Add `BeforeFilesResolve` event
If a lot of big files are uploaded at once using plupload, we would like to display some feedback to the user
while those images are being Filtered and Resolved. Otherwise, it can look like the UI is unresponsive or has errored.
See the following example to understand better:
https://www.dropbox.com/s/16675t49kvgrqlk/time%20span%20between%20file%20picker%20and%20upload%20animation%20too%20long.mov?dl=0

This change would allow us to listen to a `BeforeFilesResolve` event that would be triggered after a user submits
files to be uploaded, but before those files have been filtered and added to the queue (which is when the `FilesAdded` event is triggered).